[win32gdk] Port to Win64
Replace several `long` by `__int64` and/or `intptr_t`, `Long_t` by `Longptr_t`, `int` by `size_t`, and fix pointer formatting

+++ graf2d/win32gdk/gdk/src/glib/gutils.h
@@ -230,7 +230,7 @@ g_bit_nth_lsf (gulong mask,
   do
     {
       nth_bit++;
-      if (mask & (1 << (gulong) nth_bit))
+      if (mask & ((gulong)1 << (gulong) nth_bit))
 	return nth_bit;
     }
   while (nth_bit < 32);
@@ -245,7 +245,7 @@ g_bit_nth_msf (gulong mask,
   do
     {
       nth_bit--;
-      if (mask & (1 << (gulong) nth_bit))
+      if (mask & ((gulong)1 << (gulong) nth_bit))
 	return nth_bit;
     }
   while (nth_bit > 0);
